initial complete create scenario flow from panel
